feat(dbt): migrate all models from DuckDB to BigQuery dialect
- profiles.yml: DuckDB/MotherDuck to BigQuery (oauth dev, service-account prod)
- dbt_project.yml: varchar to STRING, int to INT64, numeric to NUMERIC seed types
- sources: spotrac.yml schema main to staging, raw.yml adds 4 missing sources
- staging (6 models): INT to INT64, DECIMAL to NUMERIC, SAFE_CAST, fix CURRENT_YEAR
- marts (6 models): ::numeric to CAST(AS NUMERIC), BigQuery PERCENTILE_CONT syntax
- Fixed SQL bugs: WHERE rn=1 on window alias, column mismatch, MAKE_DATE to DATE()

+++ dbt/models/marts/fct_player_dead_money_features.sql
@@ -31,10 +31,10 @@ rosters AS (
         player_name,
         team,
         year,
-        CAST(age AS INT) as age_at_year,
-        CAST(games_played AS INT) as games_played,
-        CAST(approximate_value AS DECIMAL(10,2)) as performance_av,
-        CAST(years_experience AS INT) as years_experience
+        CAST(age AS INT64) as age_at_year,
+        CAST(games_played AS INT64) as games_played,
+        CAST(approximate_value AS NUMERIC) as performance_av,
+        CAST(years_experience AS INT64) as years_experience
     FROM {{ ref('stg_player_rosters') }}  -- Will be created as companion model
     WHERE age IS NOT NULL
 ),
